Add files via upload
A/B tests are very commonly performed by data analysts and data scientists. It is important
that you get some practice working with the difficulties of these
For this project, you will be working to understand the results of an A/B test run by an ecommerce
website. Your goal is to work through this notebook to help the company understand
if they should implement the new page, keep the old page, or perhaps run the experiment longer
to make their decision.